src/main/java/com/forum/web/controller/LoginController.java
package com.forum.web.controller;

import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.servlet.ModelAndView;

import java.util.Map;

@Controller
public class LoginController {

    @RequestMapping(value = "/login")
    public ModelAndView loginView() {
        return new ModelAndView("login");
    }

    @RequestMapping(value = "/login", method= RequestMethod.POST)
    public ModelAndView loginStatusView(Map<String, String> params) {
        ModelAndView modelandview = new ModelAndView("login");
        return modelandview;
    }
}


